Don't make NodeText a block element to apply ellipsis in Layout panel
In bug 1109177 I cleaned up the display of nodes listed in the Layout
panel by making them not wrap, and have an ellipsis when they were too
long.
This caused a visual regression in the breadcrumbs of the Elements tree.
The nodes there use the same component and they were affected by my
CSS change.
Namely, the focus outline became taller and got cropped.

In this change, I'm just reverting the change of CSS to NodeText.
Instead, I'm moving this to the layoutPane.css, so it now only applies
to nodes in the Layout pane, and not the breadcrumbs anymore.

Before: https://imgur.com/CTGA6SO.jpg
After: https://imgur.com/sFfUQLk.jpg

Bug: 1124250
Change-Id: I9433cf0187be17a0044766472236470bb1be4b79
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2389035
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Michael Liao <michael.liao@microsoft.com>