Adds Puppeteer support to frontend
This CL adds Puppeteer as frontend dependency. There are no call sites
that use this, and in this first CL only the Puppeteer Connection class
is exposed for use. As Puppeteer is agnostified to its environment, more
of it can be made available for use within the DevTools frontend
codebase.

Bug: 1107392
Change-Id: Ie29907af389eddb2e3a7bd260b64237529a9aeba
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2354098
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>