Fix unreachable nodes in Elements pane when scrolled horizontally
In this recent CL: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2366972/
I made a change so that the Elements pane would scroll horizontally when
it became too narrow to accomodate the displayed nodes (instead of
forcing them to wrap as veyr narrow columns of characters).

This is a follow-up fix to this change.
I realized that when the pane was scrolled to the right, then hovering
over nodes would not display the usual hovered background color or
highlight the corresponding node in the page.

The first part of the fix was to make sure long attributes could break
anywhere they needed. This is useful since a lot of pages have <a>
or <script> nodes that have very long url attributes.

The second part was making the nodes inline-block display so they could
naturally expand the whole width of the scrollable container instead of
being constrained to 100% (which does not covers the overflow area).

Before: https://imgur.com/O6ya2Td.gif
After: https://imgur.com/U4sWMhj.gif

Bug: 880265
Change-Id: I4a4900052e25839f015e4fee3d20be500d8a0de4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2393465
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Kalon Hinds <kahinds@microsoft.com>