Fix jumping horizontal scrollbar in the Elements pane
The recent fixes done in the Elements pane to force a horizontal
scrollbar when there isn't enough space have introduced a regression
whereby when hovering over nodes in the Elements tree, a scrollbar can
sometimes appear, and disappear as soon as you stop hovering.

Those recent fixes are:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2393465/
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2366972/

The reason the jumpy scrollbar occurs is that when a node is hovered,
instead of having its normal 12px padding left (to indent it compared to
its parent node), it gets a 5px padding left + 1px border + 6px margin.
This is so that expanded nodes can have a left border to easily see how
far down the go in the tree.

Moving from a padding to a combination of padding, margin and border
messes with how the width is calculated. And because of my previous
changes to the Elements pane, it forced a scrollbar to appear where it
should not.

My fix here is to always keep the 12px padding, and instead of using a
border to draw the indentation line, we now use a linear-gradient.
The advantage is that it doesn't mess with the width of the element.

Before: https://imgur.com/qDtSrug.gif
After: https://imgur.com/Mjl1vMz.gif

Bug: 1128873
Change-Id: Ie6952f4631e7f5842578fbe4429b66b2f3dc9111
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2414748
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Kalon Hinds <kahinds@microsoft.com>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>