Add final separator to Rendering tab
The Rendering tab is supposed to have some padding at the bottom, but a
recent upstream change [1] inadvertently removed this.

This patch adds a final transparent separator to the Rendering tab to
ensure the .chrome-select-label margin-bottom is still applied as
originally intended.

[1]: https://chromium-review.googlesource.com/c/chromium/src/+/2371203

Fixed: chromium:1129369
Change-Id: I2a27d124ac46ea4421682d794e3b2477963b565c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2421702
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Mathias Bynens <mathias@chromium.org>