Add back + button for style insertion of stylesheets
This partially reverts
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2232549
which removed the "three-dot style menu". The menu contained multiple
things, one of which was the insertion of style sheet rules in a
specific stylesheet.

While the other color-related removals were mostly unused, the insertion
of a style rule in a specific stylesheet was used more often than we
anticipated. User feedback requested the feature to be added, as the
style insertion button on the top right would require a long-click
to insert into a specific stylesheet.

The code is largely taken from the original CL, but is slightly rewritten
as most of the buttons of the menu are removed and we only need to add
one button.

R=aerotwist@chromium.org

Fixed: 1128839
Change-Id: Ib54dc3c41c983ebe4f99e03b1d4f963ad62499c5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2422952
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>