Guard against malformed URIs when creating a binding for a resource
In Automapping._createBinding, we already guard against malformed URIs
when those are snippets or file URIs, but we don't otherwise.

This can sometimes lead to silent exceptions as described in bug 1132305.
This change simply adds the same guard in place to avoid the exception.

Bug: 1132305
Change-Id: I02cb33669f3dc5adabac852e3c87dc475c6d0245
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2431808
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Vidal Diazleal <vidorteg@microsoft.com>