Changing 'request blocking' to 'network request blocking'
'request blocking' can be confusing for users using a screen reader and
no context on the current panel. e.g when the user opens the Devtools
and the first panel shown is the 'request blocking' because it was used
last before ending the previous session.

Change-Id: If8390fd8b97e75a4ccd53d068787c1fc424f1d32
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2438829
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Christy Chen <chrche@microsoft.com>
Commit-Queue: Vidal Diazleal <vidorteg@microsoft.com>