Correctly handle multiple frames with same URL in frame tree
In the frame details view the link to a document source was generated
by its URL only. It is possible to have multiple documents with the
same URL (in different frames). This CL now takes the frame into
account as well when generating the link.

Fixed: chromium:1122502
Change-Id: Ibd922c8c3fb3e5e3a2d29d8a3e69c48e0d000f72
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2438252
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>