Fix exception when snippets scripts is an empty array
At line 222, the scriptsForSourceURL(url) function call would return either an array of strings, or empty array.
The case of an empty array would cause the scripts[scripts.length - 1] to be undefined, and throws an exception of cannot read property 'scriptId' of undefined.
This CL adds an early return in that case.

Change-Id: Ia656ac4da0767adf2d68f077f8ff0985c31e3514
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2441884
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Christy Chen <chrche@microsoft.com>