Fix heap snapshot saving
The bug is caused by a change in removeURLFragment[1] that accepts only
valid URLs now. It turns out that InspectorFrontendHost can receive
a filename instead of an URL when a file is generated by DevTools. This
CL adds a try/catch around URL trimming/fragment removal part and falls
back to whatever is provided if an URL is invalid.

[1]: https://crrev.com/c/2224810

Fixed: 1131508
Change-Id: I33878bdbc29e404752d3b9685538499d929c84ce
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2440480
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>