Add text node updates to Mutation test helpers
This change refactors the Mutation observer test helpers to enable them
to track changes within text nodes, rather than only look for DOM
additions and removals.

I took the opportunity to refactor the types slightly and also added a
suite of tests for the test helpers because they are doing quite a lot
and some of the logic is a bit fiddly.

Bug: 1129881
Change-Id: Iaaa6d289f6dad9e055eff8715de4f5a21ae5a115
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2440606
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>