Fix CSS priority for generated dark-mode themes
This CL makes sure that CSS variables are correctly copied to the
generate theme files even if they are not altered. This is important
as otherwise the style (although present in the basic style sheet)
might not apply, as its priority is not high enough.

Bug: chromium:1135123
Change-Id: I8376615d9bea5d4a886cdf268e940f3494912fe9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2449597
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>