Add dedicated workers to frame tree
This CL adds dedicated web workers to the frame tree. They are added
to the tree as a child of the frame which creates them. Clicking on
a worker in the frame tree shows a rudimentary view of the web worker's
details.

Screenshot: https://i.imgur.com/heP2paY.png

Bug: chromium:1122507
Change-Id: I4ba4a6f267944a4f95b08ced738c330e615b31e3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2463347
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>