Fix type for SortableDataGrid.create
SortableDataGrid.create accepts a list of values which is used to
construct the data argument for the SortableDataGridNode constructor
by filling the constructor's parameter
  @param {?Object.<string, *>=} data
while generating the indices.

The current type required string instead of any, but some callers
are passing objects here.

Once we are done with Typescriptification (and know all caller
provided types), we should consolidate this to something narrower.

Bug: chromium:1011811, chromium:1135143
Change-Id: I6cf7a8aeaea9eab5092c103d5f2ae1c7b768365c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2466185
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>