Breadcrumbs: don't reset user scroll position on update
The breadcrumbs update method would reset a user's scroll position
before re-rendering every time. After render, it would then scroll to
the right place.

This causes a bug where the active element is continuously scrolled in
and out of view as a re-render sets the scroll to 0 before it then
scrolls the active element into view. You can see this in the bug that
the CL fixes.

The two properties `overflowing` and `userScrollPosition` are set to
`false` and `start` accordingly in the constructor, so we default to
that on a first render (which is all we can do as we don't have anything
else to go on) and then on subsequent renders we now pick up from where
we left off, rather than blindly resetting.

Bug: 1138767
Change-Id: I265e7fe2c39cb6014f216cc64c25176914af974b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2484725
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>