[cleanup] Remove string internalization from DebuggerModel
The `DebuggerModel` instances had a `_stringMap` property, which used to
hold a `Map` to internalize certain strings, namely hash and sourceURL
of parsed scripts. It's unclear why this optimization was added in the
first place, it was probably used more throughout the code to reduce
memory footprint, but in it's current form the optimization doesn't save
anything, but rather wastes memory. Also if such an optimization is
found to be useful, it should be done in V8 instead of ad-hoc for some
random strings.

Change-Id: I974714339074bd7f03ee582c20bd2e85beefd4db
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2498228
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>