Angle component style fixes
This CL adds more shadow to the clock and its hand, and uses proper
math to calculate the hand and the dials' positions instead of using
the transparent gradient trick. This makes round edges in the hand
and the dials possible.

Bug: chroimum:1141435, chromium:1126178
Change-Id: Ia61e957b7537a0b7739cc83ede75d2999a2d7106
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2494809
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Auto-Submit: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Adam Argyle <argyle@chromium.org>